requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-markdownify==0.9.3
djangorestframework==3.14.0
Markdown==3.5.1
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
webencodings==0.5.1
django-environ
whitenoise
gunicorn
